requirements.txt
# GTX 1650 Optimized Dependencies for DeepFake Detection
# Compatible with CUDA 11.8 and TensorFlow 2.13

# Core ML Framework
tensorflow-gpu==2.13.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-image==0.21.0

# Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Transformers for ViT
transformers==4.33.2
torch==2.0.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# Explainable AI
shap==0.42.1

# Performance Monitoring
nvidia-ml-py3==7.352.0
psutil==5.9.5

# Utilities
tqdm==4.66.1
ipython==8.15.0

# Optional: For better data loading
albumentations==1.3.1
